feat:update project structure and refactor code
