+++ src/App.css
@@ -41,11 +41,10 @@ button:hover {
 /* Countdown text styles */
 .countdown-text {
   display: flex;
-  justify-content: space-between;
+  justify-content: center;
   margin: 24px 0px;
   display: flex;
   flex-direction: column; /* Stacks the elements vertically */
-  justify-content: center;
   align-items: center;
   margin-top: 20px;
   font-size: 16px;
@@ -57,15 +56,26 @@ button:hover {
   font-weight: 400;
   color: #2d1d35;
   margin-bottom: 20px;
+  text-align: center;
 }
 
 /* Countdown text button styles */
-.countdown-text > button {
+.countdown-text button {
   background-color: #fff;
   border: none;
   font-weight: 500;
   text-decoration: underline;
   cursor: pointer;
+  margin-top: 10px;
+}
+.countdown-wrapper button {
+  background-color: #fff;
+  border: none;
+  font-weight: 500;
+  text-decoration: underline;
+  cursor: pointer;
+  margin-left: 5px; /* Adjust spacing between text and button */
+  padding: 0;
 }
 
 /*Styling for Register Page */
