+++ src/constants.ts
@@ -1,1 +1,2 @@
 export const RE_DIGIT = new RegExp(/^\d+$/);
+export const PHONE_NUMBER_REGEX = /^\d*$/; // Allow only digits (empty string is valid for clearing the input)
