--- src/constants.tsx
@@ -1,1 +1,0 @@
-export const PHONE_NUMBER_REGEX = /^\d*$/; // Allow only digits (empty string is valid for clearing the input)
