+++ src/pages/RegisterPage.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import WebankLogo from "../assets/Webank.png";
 import countryOptions from "../assets/countries.json";
 import parsePhoneNumberFromString from "libphonenumber-js";
-import { PHONE_NUMBER_REGEX } from "../constants.tsx";
+import { PHONE_NUMBER_REGEX } from "../constants.ts";
 
 type CountryOption = {
   value: string;
